+++ apps/wallet/src/pages/wallet-add/restore-wallet/WalletRestore.tsx
@@ -1,8 +1,7 @@
 import React, { useContext, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 
-import { WalletType } from '@minotaur-ergo/types';
-import { MAIN_NET_LABEL } from '@minotaur-ergo/utils';
+import { MAIN_NET_LABEL, WalletType } from '@minotaur-ergo/types';
 import { Button, Grid } from '@mui/material';
 import CircularProgress from '@mui/material/CircularProgress';