+++ packages/utils/src/index.ts
@@ -1,5 +1,4 @@
 export * from './array';
-export * from './const';
 export * from './enc';
 export * from './network';
 export * from './json';